feat: added `disallowedSchemaNames` argument for security reasons

+++ src/Command/Doctrine/DoctrineSchemaDropCommand.php
@@ -11,8 +11,13 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class DoctrineSchemaDropCommand extends AbstractDoctrineSchemaCommand
 {
-    public function __construct(Connection $connection)
-    {
+    /**
+     * @param string[] $disallowedSchemaNames
+     */
+    public function __construct(
+        Connection $connection,
+        private readonly array $disallowedSchemaNames = [],
+    ) {
         parent::__construct('doctrine:schema:delete', $connection);
     }
 
@@ -21,6 +26,14 @@ class DoctrineSchemaDropCommand extends AbstractDoctrineSchemaCommand
         OutputInterface $output,
     ): int {
         $schema = $this->getSchemaFromInput($input);
+
+        if (in_array($schema, $this->disallowedSchemaNames, true)) {
+            $output->writeln(
+                "<error>Command is disallowed from being called for the '$schema' schema</error>"
+            );
+
+            return Command::FAILURE;
+        }
 
         $output->writeln("<info>Drop schema '{$schema}'...<info>");
